feat: split http and wails into two entrypoints
