--- frontend/.env.example
@@ -1,2 +1,0 @@
-VITE_APP_TYPE=HTTP
-#VITE_APP_TYPE=WAILS
