+++ frontend/package.json
@@ -5,7 +5,9 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "dev:wails": "VITE_NWC_APP_TYPE=WAILS vite",
     "build": "tsc && vite build",
+    "build:wails": "tsc && VITE_NWC_APP_TYPE=WAILS vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
